add csv upload and parsing functionality
created a upload component for the admin dashboard and when the file is upload it returns the user to the dashboard. Added csv parsing to the admin api and used the fs and csv-parser the read the file. I stored the contents of the file into a object in order to help make the db queries easier.