Parse Questions coming from csv file -Ari
Questions in a csv file are formatted as Multi-choice questions with 4 options [Q.txt, A1.txt, A1,A2.txt, A2,A3.txt, A3,A4.txt, A4]. Each question is stored in a double array (questions and the question elements).
One of the A(1-4) has a * in front. When parsing, for loops are used to find the A with the * and save it as var temp which then gets add to the new question in database as Correct.